WIP move mesh rendering functions to scene
